feat(dockerfile): Example of running `Node` under `node` user
By default a container runs process as `root`.

Best practices says that your application should run under a dedicated
user with limited access.

In the case of `node` user that is defined in the `Node` container image
it is created by the image creators. A container `foo` will not
automatically have an unprivileged `foo` user.